pyproject.toml
[project]
name = "sentinel"
version = "0.1.0"
description = "Autonomous SRE agent that monitors containers, diagnoses failures, and takes action using DSPy-trained reasoning"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepeval>=1.3.0",
    "docker>=7.1.0",
    "dspy-ai>=2.5.0",
    "fastapi>=0.123.10",
    "google-adk>=0.1.5",
    "google-genai>=1.0.0",
    "mcp>=1.0.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
